fix(shared): enforce integer amounts and remove any from sync payload type
- Add .int() validation to all amount fields (session, expense, creditEntry)
- Add .int() to time rate fields (hour, half) and frame rate value
- Replace z.any() with z.unknown() in sync payload to satisfy type constraints

Fixes:
- Prevents fractional money amounts that violate integer-only constraint
- Ensures TypeScript infers unknown instead of any for sync payloads

+++ packages/shared/src/schemas/creditEntry.schema.ts
@@ -5,12 +5,12 @@ export const CreditEntrySchema = z.object({
   customerId: z.string().uuid(),
   date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
   type: z.enum(['CREDIT_GIVEN', 'PAYMENT_RECEIVED']),
-  amount: z.coerce.number().positive('Amount must be greater than 0'),
+  amount: z.coerce.number().int().positive('Amount must be greater than 0'),
   updatedAt: z.string().datetime().optional(),
 });
 export type CreditEntry = z.infer<typeof CreditEntrySchema>;
 
 export const CreditDraftSchema = z.object({
-  amount: z.coerce.number().positive('Amount must be greater than 0'),
+  amount: z.coerce.number().int().positive('Amount must be greater than 0'),
 });
 export type CreditDraft = z.infer<typeof CreditDraftSchema>;
